// types.ts
export interface ReportData {
  namaPertandingan: string;
  guruBertanggungjawab: string;
  disahkanOleh: string;
  jawatanDisahkanOleh: string;
  tarikh: string;
  masa: string;
  tempat: string;
  sasaran: string;
  t4L: string;
  t4P: string;
  t5L: string;
  t5P: string;
  t6L: string;
  t6P: string;
  kekuatan: string;
  kelemahan: string;
  cadangan: string;
  images: string[];
}